requirements.txt

google-cloud==0.34.0
google-cloud-bigquery==1.28.0
google-cloud-storage==1.31.2
pandas_gbq==0.13.2
argparse
tqdm==4.50.0
